Bug 486 - Modified stream structs to also be 32 bit wide.
git-svn-id: http://svn.zoneminder.com/svn/zm/trunk@2594 e3e1d417-86f3-4887-817a-d78f3d33393f

+++ src/zm_stream.h
@@ -47,12 +47,12 @@ protected:
 
 protected:
     typedef struct {
-        long msg_type;
+        int msg_type;
         char msg_data[16];
     } CmdMsg;
 
     typedef struct {
-        long msg_type;
+        int msg_type;
         char msg_data[256];
     } DataMsg;
